Tests for booking have been moved to separate file. Added initialization startDate field for datePicker.
